tensorflow fixes and implementations
Summary: still finishing this off for the notebooks, but all tests are passing and everything should be documented. I also just spent the last hour fighting with trying to get it up so I've simplified the test cases for pytorch datasets to make those go much faster

Reviewers: mgoin, tuan, alexm, ben, kevinaer

Reviewed By: mgoin

Subscribers: #core, #python_ml

Differential Revision: https://phabricator.neuralmagic.com:8021/D1732

+++ neuralmagicML/pytorch/nn/fatrelu.py
@@ -80,7 +80,7 @@ def fat_pw_relu(
     out = fat_relu(tens, x_offset, inplace)
 
     # calculate the compression region values
-    comp_mask = ((tens < threshold) * tens > x_offset).float()
+    comp_mask = ((tens < threshold).float() * tens > x_offset).float()
     comp_tens = compression * (out - x_offset)
 
     # reassign the compression values in the output
